+++ strings/lower_case.go
@@ -1,0 +1,10 @@
+package strings
+
+import "strings"
+
+// Converts string, as space separated words, to lower case.
+//
+// Complexity: O(n)
+func LowerCase(input string) string {
+	return strings.ReplaceAll(CustomDelimitedCase(input, '-', "", false), "-", " ")
+}